+++ tag-20180828/ChangeLog.md
@@ -1,3 +1,11 @@
+2024-03-14 Version: 1.0.2
+- Update API DisablePolicyType: add param OpenType.
+- Update API EnablePolicyType: add param OpenType.
+- Update API GetPolicyEnableStatus: add param OpenType.
+- Update API ListTargetsForPolicy: update response param.
+- Update API UntagResources: update param TagKey.
+
+
 2023-12-08 Version: 1.0.1
 - Generated python2 2018-08-28 for Tag.